// js/html.d.ts
export type Page = IndexCtx | AllSymbolsCtx | SymbolPageCtx | Redirect;

export interface PageBase {
  kind: "IndexCtx" | "AllSymbolsCtx" | "SymbolPageCtx";
  html_head_ctx: HtmlHeadCtx;
  disable_search: boolean;
  categories_panel: CategoriesPanelCtx | null;
  breadcrumbs_ctx: BreadcrumbsCtx;
}

export interface IndexCtx extends PageBase {
  kind: "IndexCtx";
  overview: SymbolContentCtx | null;
  module_doc: ModuleDocCtx | null;
  usage: UsagesCtx | null;
  toc_ctx: ToCCtx;
}

export interface AllSymbolsCtx extends PageBase {
  kind: "AllSymbolsCtx";
  content: SymbolContentCtx;
}

export interface SymbolPageCtx extends PageBase {
  kind: "SymbolPageCtx";
  symbol_group_ctx: SymbolGroupCtx;
  toc_ctx: ToCCtx;
}

export interface Redirect {
  kind: "redirect";
  path: string;
}

export interface ToCCtx {
  usages: UsagesCtx | null;
  top_symbols: TopSymbolsCtx | null;
  document_navigation_str: string | null;
  document_navigation: ToCEntry[];
}

export interface ToCEntry {
  level: number;
  content: string;
  anchor: string;
}

export interface UsagesCtx {
  usages: UsageCtx[];
  composed: boolean;
}
// ...
export interface ModuleDocCtx {
  deprecated: string | null;
  sections: SymbolContentCtx;
}

export interface DocNodeKindCtx {
  kind: string;
  char: string;
  title: string;
  title_lowercase: string;
  title_plural: string;
}

export interface SymbolContentCtx {
  id: string;
  docs: string | null;
  sections: SectionCtx[];
}

export interface SectionCtx {
  header: SectionHeaderCtx;
  content: SectionContentCtx;
}

export interface SectionHeaderCtx {
  title: string;
  anchor: AnchorCtx;
  href: string | null;
  doc: string | null;
}

export interface AnchorCtx {
  id: string;
}

export interface SymbolGroupCtx {
  name: string;
  symbols: SymbolCtx[];
}